+++ src/pdm/cli/commands/run.py
@@ -22,7 +22,7 @@ from pdm.cli.utils import check_project_file
 from pdm.exceptions import PdmUsageError
 from pdm.project import Project
 from pdm.signals import pdm_signals
-from pdm.utils import expand_env_vars, is_path_relative_to
+from pdm.utils import deprecation_warning, expand_env_vars, is_path_relative_to
 
 if TYPE_CHECKING:
     from typing import Any, Callable, Iterator, TypedDict
@@ -39,11 +39,18 @@ if TYPE_CHECKING:
         working_dir: str
 
 
-def exec_opts(*options: TaskOptions | None) -> dict[str, Any]:
-    return dict(
-        env={k: v for opts in options if opts for k, v in opts.get("env", {}).items()} or None,
-        **{k: v for opts in options if opts for k, v in opts.items() if k not in ("env", "help")},
+def merge_options(*options: TaskOptions | None) -> TaskOptions:
+    """Merge multiple options dicts. For the same key, the last one wins."""
+    return cast(
+        "TaskOptions",
+        {
+            "env": {k: v for opts in options if opts for k, v in opts.get("env", {}).items()},
+            **{k: v for opts in options if opts for k, v in opts.items() if k not in ("env", "help", "keep_going")},
+        },
     )
+
+
+exec_opts = merge_options  # Alias for merge_options
 
 
 RE_ARGS_PLACEHOLDER = re.compile(r"\{args(?::(?P<default>[^}]*))?\}")
@@ -117,6 +124,7 @@ class TaskRunner:
         self.hooks = hooks
 
     def get_task(self, script_name: str) -> Task | None:
+        """Get the task with the given name. Return None if not found."""
         if script_name not in self.project.scripts:
             return None
         script = cast("str | Sequence[str] | Mapping[str,Any]", self.project.scripts[script_name])
@@ -233,6 +241,14 @@ class TaskRunner:
     def run_task(
         self, task: Task, args: Sequence[str] = (), opts: TaskOptions | None = None, seen: set[str] | None = None
     ) -> int:
+        """Run the named task with the given arguments.
+
+        Args:
+            task: The task to run
+            args: The extra arguments passed to the task
+            opts: The options passed from parent if any
+            seen: The set of seen tasks to prevent recursive calls
+        """
         kind, _, value, options = task
         shell = False
         if kind == "cmd":
@@ -283,13 +299,13 @@ class TaskRunner:
                 split = shlex.split(script)
                 cmd = split[0]
                 subargs = split[1:] + ([] if should_interpolate else args)
-                code = self.run(cmd, subargs, options, chdir=True, seen=seen)
+                code = self.run(cmd, subargs, merge_options(options, opts), chdir=True, seen=seen)
                 if code != 0:
                     if not keep_going:
                         return code
                     composite_code = code
             return composite_code
-        return self._run_process(args, chdir=True, shell=shell, **exec_opts(self.global_options, options, opts))
+        return self._run_process(args, chdir=True, shell=shell, **merge_options(self.global_options, options, opts))  # type: ignore[misc]
 
     def run(
         self,
@@ -299,6 +315,7 @@ class TaskRunner:
         chdir: bool = False,
         seen: set[str] | None = None,
     ) -> int:
+        """Run a command or script with the given arguments."""
         if command in self.hooks.skip:
             return 0
         if seen is None:
@@ -308,8 +325,7 @@ class TaskRunner:
             if task.kind == "composite":
                 if command in seen:
                     raise PdmUsageError(f"Script {command} is recursive.")
-                seen = seen.copy()
-                seen.add(command)
+                seen = {command, *seen}
 
             self.hooks.try_emit("pre_script", script=command, args=args)
             pre_task = self.get_task(f"pre_{command}")
@@ -326,7 +342,7 @@ class TaskRunner:
             self.hooks.try_emit("post_script", script=command, args=args)
             return code
         else:
-            return self._run_process([command, *args], chdir=chdir, **exec_opts(self.global_options, opts))
+            return self._run_process([command, *args], chdir=chdir, **merge_options(self.global_options, opts))  # type: ignore[misc]
 
     def show_list(self) -> None:
         if not self.project.scripts:
@@ -378,7 +394,6 @@ def _fix_env_file(data: dict[str, Any]) -> dict[str, Any]:
 class Command(BaseCommand):
     """Run commands or scripts with local packages loaded"""
 
-    runner_cls: type[TaskRunner] = TaskRunner
     arguments = (*BaseCommand.arguments, skip_option, venv_option)
 
     def add_arguments(self, parser: argparse.ArgumentParser) -> None:
@@ -409,10 +424,16 @@ class Command(BaseCommand):
             help="Arguments that will be passed to the command",
         )
 
+    def get_runner(self, project: Project, hooks: HookManager, options: argparse.Namespace) -> TaskRunner:
+        if (runner := getattr(self, "runner_cls", None)) is not None:  # pragma: no cover
+            deprecation_warning("runner_cls attribute is deprecated, use get_runner method instead.")
+            return cast("type[TaskRunner]", runner)(project, hooks)
+        return TaskRunner(project, hooks)
+
     def handle(self, project: Project, options: argparse.Namespace) -> None:
         check_project_file(project)
         hooks = HookManager(project, options.skip)
-        runner = self.runner_cls(project, hooks=hooks)
+        runner = self.get_runner(project, hooks, options)
         if options.list:
             return runner.show_list()
         if options.json: