+++ src/pdm/cli/commands/venv/activate.py
@@ -45,7 +45,7 @@ class ActivateCommand(BaseCommand):
         try:
             shell, _ = shellingham.detect_shell()
         except shellingham.ShellDetectionFailure:
-            shell = None
+            shell = ""
         if shell == "fish":
             command, filename = "source", "activate.fish"
         elif shell == "csh":
@@ -57,7 +57,13 @@ class ActivateCommand(BaseCommand):
         activate_script = venv.interpreter.with_name(filename)
         if activate_script.exists():
             if platform.system() == "Windows":
-                return f"{shlex.quote(str(activate_script))}"
-            return f"{command} {shlex.quote(str(activate_script))}"
+                return f"{self.quote(str(activate_script), shell)}"
+            return f"{command} {self.quote(str(activate_script), shell)}"
         # Conda backed virtualenvs don't have activate scripts
-        return f"conda activate {shlex.quote(str(venv.root))}"
+        return f"conda activate {self.quote(str(venv.root), shell)}"
+
+    @staticmethod
+    def quote(command: str, shell: str) -> str:
+        if shell in ["powershell", "pwsh"]:
+            return "'{}'".format(command.replace("'", "''"))
+        return shlex.quote(command)