+++ src/pdm/project/core.py
@@ -17,7 +17,7 @@ from tomlkit.items import Array
 from pdm import termui
 from pdm._types import RepositoryConfig
 from pdm.exceptions import NoPythonVersion, PdmUsageError, ProjectError
-from pdm.models.backends import BuildBackend, get_backend_by_spec
+from pdm.models.backends import DEFAULT_BACKEND, BuildBackend, get_backend_by_spec
 from pdm.models.caches import PackageCache
 from pdm.models.python import PythonInfo
 from pdm.models.repositories import BaseRepository, LockedRepository
@@ -398,7 +398,7 @@ class Project:
         for source in result.values():
             if not source.url:
                 continue
-            source.url = expand_env_vars_in_auth(source.url)
+            source.url = DEFAULT_BACKEND(self.root).expand_line(expand_env_vars_in_auth(source.url))
             sources.append(source)
         return sources
 
@@ -570,6 +570,7 @@ class Project:
         to_group: str = "default",
         dev: bool = False,
         show_message: bool = True,
+        write: bool = True,
     ) -> None:
         deps, setter = self.use_pyproject_dependencies(to_group, dev)
         for _, dep in requirements.items():
@@ -583,7 +584,8 @@ class Project:
             else:
                 deps[matched_index] = req
         setter(cast(Array, deps).multiline(True))
-        self.pyproject.write(show_message)
+        if write:
+            self.pyproject.write(show_message)
 
     def init_global_project(self) -> None:
         if not self.is_global or not self.pyproject.empty():